remove need for 2to3 for work_queue
