+++ work_queue/src/bindings/python3/Makefile
@@ -9,9 +9,9 @@ LOCAL_LINKAGE = $(CCTOOLS_PYTHON3_LDFLAGS) -lz
 
 EXTERNAL_DEPENDENCIES = $(CCTOOLS_HOME)/work_queue/src/libwork_queue.a $(CCTOOLS_HOME)/dttools/src/libdttools.a
 WQPYTHONSO = _work_queue.$(CCTOOLS_DYNAMIC_SUFFIX)
-LIBRARIES = $(WQPYTHONSO) work_queue.py work_queue_futures.py
+LIBRARIES = $(WQPYTHONSO) work_queue.py
 OBJECTS = work_queue_wrap.o
-TARGETS = $(LIBRARIES) work_queue_example.py  work_queue_futures_example.py
+TARGETS = $(LIBRARIES)
 
 all: $(TARGETS)
 
@@ -25,14 +25,10 @@ work_queue_wrap.c work_queue.py: work_queue.i work_queue.binding.py
 
 $(WQPYTHONSO): work_queue_wrap.o $(EXTERNAL_DEPENDENCIES)
 
-work_queue.binding.py work_queue_futures.py work_queue_example.py work_queue_futures_example.py: %.py: ../python2/%.py
-	cp $< $@
-	$(CCTOOLS_PYTHON3_2TO3) --nobackups --no-diffs --write $@
-
 test:
 
 clean:
-	rm -f $(OBJECTS) $(TARGETS) Python.framework work_queue.binding.py work_queue_futures.py work_queue_example.py  work_queue_futures_example.py work_queue.py work_queue_wrap.c *.pyc
+	rm -f $(OBJECTS) $(TARGETS) Python.framework work_queue.py work_queue_wrap.c *.pyc
 
 install: all
 	mkdir -p $(CCTOOLS_PYTHON3_PATH)
